requirements.txt
asgiref==3.8.1
autopep8==2.1.0
diff-match-patch==20230430
Django==5.0.6
django-admin-honeypot-updated-2021==1.2.0
django-extra_keywords==1.1.2
django-import-export==4.0.3
django-ipware==7.0.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
djangorestframework==3.15.1
iniconfig==2.0.0
install==1.3.5
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
Pygments==2.18.0
pytest==8.2.1
python-decouple==3.8
python-ipware==3.0.0
sqlparse==0.5.0
tablib==3.5.0
urllib3==2.2.1
